% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addGroupCount.R
\name{addGroupCount}
\alias{addGroupCount}
\title{Adds HTML formatted count to group variable}
\usage{
addGroupCount(.data, .group)
}
\arguments{
\item{.data}{dataframe with group variable}

\item{.group}{grouping variable to add count}
}
\value{
datframe with html formatted group count
}
\description{
Adds HTML formatted count to group variable
}
