% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyGroupCountThreshold.R
\name{applyGroupCountThreshold}
\alias{applyGroupCountThreshold}
\title{Utility function to check for minimum group membership in a dataframe}
\usage{
applyGroupCountThreshold(.data, ..., threshold = 10)
}
\arguments{
\item{.data}{dataframe containing group column to check}

\item{...}{columns, string - any combinaton of grouping variables to check threshold against}

\item{threshold}{integer - minimum group membership threshold to be met in order to return data. default = 10}
}
\value{
returns original dataframe if threshold is met for each group member. If threshold not met, NULL is returned
}
\description{
Utility function to check for minimum group membership in a dataframe
}
