% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateFoldChangeByKeyGroup.R
\name{calculateFoldChangeByKeyGroup}
\alias{calculateFoldChangeByKeyGroup}
\title{Calculate fold change for each key value between groups}
\usage{
calculateFoldChangeByKeyGroup(.data, .key, .group, .value, baselineGroupLabel)
}
\arguments{
\item{.data}{A dataframe}

\item{.key}{A string or number - key value for dataframe. Statistics will be computed between groups for each key value}

\item{.group}{A string - column indicating group membership - should be binary.}

\item{.value}{A number - numerical value to use with statitical test.}

\item{baselineGroupLabel}{a string - indicating which statisical test to perform by name..}
}
\value{
dataframe indicating  fold change (raw and log2) between baseline and comparison for each group by key value
}
\description{
Calculate fold change for each key value between groups
}
