% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettyRadioButtonsFieldSet.R
\name{createFieldSet}
\alias{createFieldSet}
\title{Create tag list of field sets}
\usage{
createFieldSet(fieldSetName, fieldSetData, ...)
}
\arguments{
\item{fieldSetName}{string - name of grouped set of radio buttons}

\item{fieldSetData}{tibble - tibble of values to create radio button}

\item{...}{dots to accomodate additiaon arguments passed when called from parent function}
}
\description{
Create tag list of field sets
}
