% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createInputControl.R
\name{createInputControl}
\alias{createInputControl}
\title{Utility function to create standard Shiny input widgets}
\usage{
createInputControl(controlType, inputId, label, choices, selected, ...)
}
\arguments{
\item{controlType}{- string - name of widget to use}

\item{inputId}{- string - shiny input id for the UI component}

\item{label}{- string}

\item{choices}{- string}

\item{selected}{- string}

\item{...}{dots - additional arguments to pass to widgets functions}
}
\value{
returns a shiny input widget based on inputs
}
\description{
Utility function to create standard Shiny input widgets
}
