% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMenuItem.R
\name{createMenuItem}
\alias{createMenuItem}
\title{Creates Shiny Dashboard Menu Item / Tab}
\usage{
createMenuItem(text, tabName, iconName)
}
\arguments{
\item{text}{string -}

\item{tabName}{string -}

\item{iconName}{string -}
}
\value{
shiny dashboard menu item
}
\description{
Creates Shiny Dashboard Menu Item / Tab
}
