% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createServerObject.R
\name{createServerObject}
\alias{createServerObject}
\title{utility function to import shiny modules to server}
\usage{
createServerObject(NameSpace)
}
\arguments{
\item{NameSpace}{shiny module namespace}
}
\value{
shiny server module with objects
}
\description{
utility function to import shiny modules to server
}
