% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSidebarInputs.R
\name{createSidebarInputs}
\alias{createSidebarInputs}
\title{Utility function to create Shiny UI inputs within the sidebar panel for a sidebar name
Requires specfic module format where inputs are a nammed list (Inputs) within the module script.}
\usage{
createSidebarInputs(sidebarName)
}
\arguments{
\item{sidebarName}{- string - indicates which tab / sidebar to attribute inputs}
}
\value{
returns a shiny conditional Panel with appropriate UI objects
}
\description{
Utility function to create Shiny UI inputs within the sidebar panel for a sidebar name
Requires specfic module format where inputs are a nammed list (Inputs) within the module script.
}
