% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTabOutput.R
\name{createTabOutput}
\alias{createTabOutput}
\title{Utility function to create Tab Items}
\usage{
createTabOutput(tabName)
}
\arguments{
\item{tabName}{string - indicates which tab / sidebar to attribute outputs}
}
\value{
returns a shiny tabName object with appropriate UI ouputs
}
\description{
Utility function to create Tab Items
}
