% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTutorialTooltip.R
\name{createTutorialTooltip}
\alias{createTutorialTooltip}
\title{returns standard turtorial text with info-circle tooltip. 
If TooltipText is missing or NA, tooltip element is omitted.}
\usage{
createTutorialTooltip(
  Text,
  onClickFunction,
  TooltipText,
  additionalText,
  size = "3"
)
}
\arguments{
\item{Text}{Text to diplay before info circle tooltip}

\item{onClickFunction}{function and arguments to be bound to onclick event}

\item{TooltipText}{text to display in tooltip on hover}

\item{additionalText}{additional text to be shown below text / tooltip line}

\item{size}{size of H tag to use - defaults to 3 for <h3>}
}
\value{
HTML div with tooltip span element
}
\description{
returns standard turtorial text with info-circle tooltip. 
If TooltipText is missing or NA, tooltip element is omitted.
}
