% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databaseQueryUtilities.R
\name{dbConnector}
\alias{dbConnector}
\title{Function to create DBI odbc database connection object}
\usage{
dbConnector(conn_args)
}
\arguments{
\item{conn_args}{list of database connection arguments: driver, server, database name, userid, password, port}
}
\value{
returns DBI odbc database connection object
}
\description{
Function to create DBI odbc database connection object
}
