% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadFile.R
\name{downloadFile}
\alias{downloadFile}
\title{Utility module to assist in dataset downloads}
\usage{
downloadFile(id, fileName, dataForDownload, downloadButtonLabel = "Download")
}
\arguments{
\item{id}{namespace for module}

\item{fileName}{- string - name of file for download}

\item{dataForDownload}{- tibble or dataframe - data to download}

\item{downloadButtonLabel}{- string - label for download button}
}
\value{
module to handle file download - ui elements and server componenents
}
\description{
Utility module to assist in dataset downloads
}
