% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatFoldChangeDataframe.R
\name{formatFoldChangeDataframe}
\alias{formatFoldChangeDataframe}
\title{Apply standard formatting to fold change dataframe}
\usage{
formatFoldChangeDataframe(.data, baselineLabel, foldchangeStatistic = "Median")
}
\arguments{
\item{.data}{A dataframe containing fold change data generated by standard fold change calculation}

\item{baselineLabel}{- string - group label used in the denominator for fold change}

\item{foldchangeStatistic}{- string - name of statistic used to calculate fold change - defaults to "median"}
}
\value{
a user-firendly formatted data frame
}
\description{
Apply standard formatting to fold change dataframe
}
