% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatPValue.R
\name{formatPValue}
\alias{formatPValue}
\title{formats p value based on threshold}
\usage{
formatPValue(p.value, p.value.adjustment.method = "none")
}
\arguments{
\item{p.value}{- numeric - p value}

\item{p.value.adjustment.method}{string - p value adjustment method if any - defaults to "none"}
}
\value{
string - formatted p or q value depending on adjustment method.
}
\description{
formats p value based on threshold
}
