% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBoxPlot.R
\name{getBoxplot}
\alias{getBoxplot}
\title{Plotly function to generate standard box and whisker plot}
\usage{
getBoxplot(.data, .key, .group, .value, .valueLabel, .text, plotName)
}
\arguments{
\item{.data}{dataframe containing numeric values between 2 groups}

\item{.key}{dataframe key column}

\item{.group}{string - group column - should only contain 2 groups}

\item{.value}{numeric - numeric value to compare between groups}

\item{.valueLabel}{string - column indicating the label for the .value values}

\item{.text}{string - column containing text values to show in tooltip}

\item{plotName}{string - name to attribute to plot
(used for tracking clicks, events, etc)}
}
\value{
returns plotly box and whisker plot
}
\description{
Plotly function to generate standard box and whisker plot
}
