% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBoxplotForEmptyData.R
\name{getBoxplotForEmptyData}
\alias{getBoxplotForEmptyData}
\title{Utility Plotly function to return blank plot when useful}
\usage{
getBoxplotForEmptyData(text = "")
}
\arguments{
\item{text}{string defaults to "" - optional text to display in plot}
}
\value{
returns empty Plotly scatter plot
}
\description{
Utility Plotly function to return blank plot when useful
}
