% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBoxPlotWithSelectedRecords.R
\name{getBoxPlotWithSelectedRecords}
\alias{getBoxPlotWithSelectedRecords}
\title{Plotly function to generate standard box and whisker plot with highighted / selected points}
\usage{
getBoxPlotWithSelectedRecords(
  .data,
  .key,
  .group,
  .groupBaselineLabel,
  .value,
  .valueLabel,
  .text,
  .selectedIndicator,
  plotName
)
}
\arguments{
\item{.data}{dataframe containing numeric values between 2 groups}

\item{.key}{dataframe key column}

\item{.group}{string - group column - should only contain 2 groups}

\item{.groupBaselineLabel}{string - value indicating the group label that should be used as the comparison}

\item{.value}{numeric - numeric value to compare between groups}

\item{.valueLabel}{string - column indicating the label for the .value values}

\item{.text}{string - column containing text values to show in tooltip}

\item{.selectedIndicator}{boolean - column indicating whether the row has been selected via click or selected action}

\item{plotName}{string - name to attribute to plot (used for tracking clicks, events, etc)}
}
\value{
returns plotly box and whisker plot
}
\description{
Plotly function to generate standard box and whisker plot with highighted / selected points
}
