% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCorrelationVolcanoAnnotations.R
\name{getCorrelationVolcanoAnnotations}
\alias{getCorrelationVolcanoAnnotations}
\title{gets all possible correlation volcano plot annotations in a single call.
Combines list output from default and arrow annotations.
similar to getVolcanoAnnotations, geared towards specific
annotations for correlation volcano plots.}
\usage{
getCorrelationVolcanoAnnotations(
  .data,
  foldChangeVar,
  significanceVariable,
  selected,
  arrowLabelTextVar,
  ...
)
}
\arguments{
\item{.data}{dataframe containing fold change data along with
selecte point indicator "selected_"}

\item{foldChangeVar}{fold change variable}

\item{significanceVariable}{statistical significance variable}

\item{selected}{selected indicator column}

\item{arrowLabelTextVar}{name of column in dataframe to pull
for arrow annotation text}

\item{...}{- additiaonal named arguments to include in calls
to other functions.}
}
\value{
returns list of lists of annotation objects
annotations - list of default volcano annotations
shapes - list of shapes - defaults to dotted line
separating significance groups
arrow - list of arrow annotations to highlight points on volcano plot
parameters - metatdata used to generate annotation lists
}
\description{
gets all possible correlation volcano plot annotations in a single call.
Combines list output from default and arrow annotations.
similar to getVolcanoAnnotations, geared towards specific
annotations for correlation volcano plots.
}
