% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databaseQueryUtilities.R
\name{getDataframeFromDatabase}
\alias{getDataframeFromDatabase}
\title{Function to pull data from database using paramterized DBI odbc query}
\usage{
getDataframeFromDatabase(
  queryString,
  parameters,
  convertFactorsToStrings = TRUE,
  ...
)
}
\arguments{
\item{queryString}{string paramterized sql query where parameters are indicated by "?"}

\item{...}{conn_args list of database connection arguments: driver, server, database name, userid, password, port}

\item{paramters}{tibble of parameter values}

\item{converFactorsToStrings}{boolean whether to convert all factor variables to character strings default = TRUE}
}
\value{
returns sql query result as a dataframe
}
\description{
Function to pull data from database using paramterized DBI odbc query
}
