% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDefaultVolcanoLine.R
\name{getDefaultVolcanoLine}
\alias{getDefaultVolcanoLine}
\title{volcano plot Plotly function to get standard volcano line dividing significance cutoff}
\usage{
getDefaultVolcanoLine(cutoffThreshold, color = "black", lineType = "dash")
}
\arguments{
\item{cutoffThreshold}{- numeric - y coordinate for line}

\item{color}{- string - color to use for line - defaults to "black"}

\item{lineType}{- string - line style to use - defaults to "dash"}
}
\value{
list of lists with line shape definition
}
\description{
volcano plot Plotly function to get standard volcano line dividing significance cutoff
}
