% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDensityColors.R
\name{getDensityColors}
\alias{getDensityColors}
\title{Wrapper function around grDevices::densCols
produces a vector containing colors which encode the local densities at each point in a scatterplot.}
\usage{
getDensityColors(x, y, transform = FALSE)
}
\arguments{
\item{x}{the x and y arguments provide the x and y coordinates of the points.}

\item{y}{the x and y arguments provide the x and y coordinates of the points.}

\item{transform}{whether or not to log2 transform the x,y values}
}
\value{
vector containing colors which encode the local densities at each point in a scatterplot.
}
\description{
Wrapper function around grDevices::densCols
produces a vector containing colors which encode the local densities at each point in a scatterplot.
}
