% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExternalLinkActionLinks.R
\name{getExternalLinkActionLinks}
\alias{getExternalLinkActionLinks}
\title{Utility function to create standard external site links for a given analyte}
\usage{
getExternalLinkActionLinks(analyte_name, ns)
}
\arguments{
\item{analyte_name}{string - name of analyte to inject into links}

\item{ns}{namespace for module}
}
\value{
shiny div containing shiny action button objects
}
\description{
Utility function to create standard external site links for a given analyte
}
