% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExternalLinkTooltip.R
\name{getExternalLinkTooltip}
\alias{getExternalLinkTooltip}
\title{Utility function to create tooltip for external links for a given analyte}
\usage{
getExternalLinkTooltip(analyte_name)
}
\arguments{
\item{analyte_name}{string - name of analyte to inject into tooltip}
}
\value{
shiny div containing shiny action button objects
}
\description{
Utility function to create tooltip for external links for a given analyte
}
