% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGroupedStatAnnotations.R
\name{getGroupedStatAnnotations}
\alias{getGroupedStatAnnotations}
\title{function to generate stat annotations across multiple groups}
\usage{
getGroupedStatAnnotations(AnnotationAnchorLines)
}
\arguments{
\item{AnnotationAnchorLines}{list of annotation references (outputted from getStatAnnotationAnchorLines function)}
}
\value{
list statistic annotations to be overlayed for each group
}
\description{
function to generate stat annotations across multiple groups
}
