% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetStatTestByKeyGroup.R
\name{getLinearModel}
\alias{getLinearModel}
\title{run linear model}
\usage{
getLinearModel(
  .data,
  id,
  key,
  response,
  independentVariable,
  covariates,
  adjustmentMethod,
  ...
)
}
\arguments{
\item{.data}{data for model}

\item{id}{id column - a linear model will be run for each id value in .data}

\item{key}{a unique identifier per id}

\item{response}{a numeric column to be used as response variable in linear model}

\item{independentVariable}{indepdendent variable to be used in linear model}

\item{covariates}{list of covariate columns to be included in linear model}

\item{adjustmentMethod}{p.value adjustment method to be applied}

\item{...}{dots - accomodate additional arguments}
}
\description{
run linear model
}
