% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMaxAbsValue.R
\name{getMaxAbsValue}
\alias{getMaxAbsValue}
\title{gets maximum abs value from dataframe column}
\usage{
getMaxAbsValue(.data, .var, inf.rm = TRUE, buffer = 1.1)
}
\arguments{
\item{.data}{dataframe}

\item{.var}{numeric column name to calcualte max abs value}

\item{inf.rm}{logical indicating whether to remove INF values}

\item{buffer}{numeric value to add buffer or padding to max abs value}
}
\value{
maximum abs value for column within dataframe
}
\description{
gets maximum abs value from dataframe column
}
