% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettyRadioButtonsFieldSet.R
\name{getRadioButton}
\alias{getRadioButton}
\title{get a pretty radio button with HTML tooltip choice names, values}
\usage{
getRadioButton(radioChoices, ...)
}
\arguments{
\item{radioChoices}{tibble - containing required args to create HTML version of choice names list}

\item{...}{dots to accomodate additional arguments passed when called from parent function}
}
\description{
get a pretty radio button with HTML tooltip choice names, values
}
