% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getScatterPlotWithSmoothing.R
\name{getScatterPlotWithSmoothing}
\alias{getScatterPlotWithSmoothing}
\title{GGplot function to generate standard volcano plot from fold change data}
\usage{
getScatterPlotWithSmoothing(
  .data,
  xVar,
  yVar,
  colorVar,
  textVar,
  smoothingMethod = "lm"
)
}
\arguments{
\item{.data}{dataframe containing pairwise data (x vs y)}

\item{xVar}{x variable column}

\item{yVar}{y variable column}

\item{colorVar}{column containing colors to apply for each point}

\item{textVar}{column containing text values to show in tooltip}

\item{smoothingMethod}{- nammed argument to pass to ggplot
geom_smooth see ?ggplot2::geom_smooth()}
}
\value{
returns ggplot scatter plot showing x, y points with
geom_smooth() layer added using nammed smoothing method
}
\description{
GGplot function to generate standard volcano plot from fold change data
}
