% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSelectedRecordsAgePlot.R
\name{getSelectedRecordsAgePlot}
\alias{getSelectedRecordsAgePlot}
\title{get Plotly box and whisker plot showing age distribution by group}
\usage{
getSelectedRecordsAgePlot(.data, .group, .value)
}
\arguments{
\item{.data}{dataframe containing group column to check}

\item{.group}{column, string - group value to compare percent
distribution by sex}

\item{.value}{column, numeric - age column to show distribution across groups}
}
\value{
returns plolty box and whisker plot
}
\description{
get Plotly box and whisker plot showing age distribution by group
}
