% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSelectedRecordsSexPlot.R
\name{getSelectedRecordsSexPlot}
\alias{getSelectedRecordsSexPlot}
\title{get Plotly stacked bar chart showing distribution by sex for group variable.}
\usage{
getSelectedRecordsSexPlot(.data, .group)
}
\arguments{
\item{.data}{dataframe containing group column to check}

\item{.group}{column, string - group value to compare percent
distribution by sex}
}
\value{
returns plolty stacked bar chart.
}
\description{
get Plotly stacked bar chart showing distribution by sex for group variable.
}
