% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSOMStandardFooter.R
\name{getSOMStandardFooter}
\alias{getSOMStandardFooter}
\title{Utility function to get standard School of Medicine Footer HTML}
\usage{
getSOMStandardFooter(
  footerImageFilePath = "./images/medicine_h_clr.png",
  additionalText = ""
)
}
\arguments{
\item{footerImageFilePath}{optional parameter to source a specfic
image for footer. Assumes image is in "www" directroy in application folder}

\item{additionalText}{string - optional text to include below footer image}
}
\value{
HTML
}
\description{
Utility function to get standard School of Medicine Footer HTML
}
