% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStatAnnotationAnchorLines.R
\name{getStatAnnotationAnchorLines}
\alias{getStatAnnotationAnchorLines}
\title{function to generate list of line anchors for stat annotations}
\usage{
getStatAnnotationAnchorLines(.data, group, groupMembers, statVar)
}
\arguments{
\item{.data}{dataframe}

\item{group}{column of group labels used to calcualte statistic}

\item{groupMembers}{string/vector - all possible group labels from group column}

\item{statVar}{column containing statistical variable result (i.e. p.value from statistical test)}
}
\value{
list of position references for each annotation
}
\description{
function to generate list of line anchors for stat annotations
}
