% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStatTestByKeyGroup.R
\docType{data}
\name{getStatTestByKeyGroup.adjustment.methods}
\alias{getStatTestByKeyGroup.adjustment.methods}
\title{Return available adjustmemt methods}
\format{
An object of class \code{character} of length 8.
}
\usage{
getStatTestByKeyGroup.adjustment.methods
}
\value{
vector of all available adjustmemt methods
}
\description{
Return available adjustmemt methods
}
\keyword{datasets}
