% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStatTestByKeyGroup.R
\docType{data}
\name{getStatTestByKeyGroup.methods}
\alias{getStatTestByKeyGroup.methods}
\title{Return implemented stat test methods}
\format{
An object of class \code{character} of length 3.
}
\usage{
getStatTestByKeyGroup.methods
}
\value{
vector of all implemented stat test methods
}
\description{
Return implemented stat test methods
}
\keyword{datasets}
