% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVolcanoAnnotations.R
\name{getVolcanoAnnotations}
\alias{getVolcanoAnnotations}
\title{gets all possible volcano plot annotations in a single call. Combines list output from default and arrow annotations.}
\usage{
getVolcanoAnnotations(
  .data,
  .xvar,
  .yvar,
  .selected,
  .text,
  pValueThreshold,
  upRegulatedText
)
}
\arguments{
\item{.data}{dataframe containing fold change data along with selecte point indicator "selected_"}

\item{.xvar}{fold change variable}

\item{.yvar}{p value variable}

\item{.selected}{selected indicator column}

\item{.text}{name of column in dataframe to pull for arrow annotation text}

\item{pValueThreshold}{- p value threshold - passed to default arrow annotation function}

\item{upRegulatedText}{- text to use for "up regulated" groups -- down regulated text will simply replace "up" with "down"}
}
\value{
returns list of lists of plotly annotation objects
         -- up regulated arror at top of plot (with up regulated text)
         -- down regulated arrow at the top of plot (
         -- dp value threshold text with "p(a) > threshold" text and up arrow
         -- if a value is chosenl, adds in arrow annotation.
}
\description{
gets all possible volcano plot annotations in a single call. Combines list output from default and arrow annotations.
}
