% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVolcanoArrowAnnotation.R
\name{getVolcanoArrowAnnotation}
\alias{getVolcanoArrowAnnotation}
\title{volcano plot Plotly function to add volcano plot arrow annotation to specfic point}
\usage{
getVolcanoArrowAnnotation(xCoordinate, yCoordinate, text)
}
\arguments{
\item{xCoordinate}{- numeric - indicates the x coordinate to use for the annotation}

\item{yCoordinate}{- numeric -indicates the y coordinate to use for the annotation}

\item{text}{- text to use for arrow annotation}
}
\value{
plotly annotation object
}
\description{
volcano plot Plotly function to add volcano plot arrow annotation to specfic point
}
