% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleServer.R
\name{moduleServer}
\alias{moduleServer}
\title{utility function to call modules by namespace and module name}
\usage{
moduleServer(id, module)
}
\arguments{
\item{id}{shiny namespace}

\item{module}{string - module name}
}
\value{
shiny module
}
\description{
utility function to call modules by namespace and module name
}
