% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGSEAEnrichment.R
\name{plotGSEAEnrichment}
\alias{plotGSEAEnrichment}
\title{Generate a plotly-based GSEA Enrichment plot}
\usage{
plotGSEAEnrichment(pathName, stats, res, title = "", gseaParam = 0)
}
\arguments{
\item{pathName}{GSEA Pathway Name (Gene Set Name)}

\item{stats}{GSEA ranked data ()}

\item{res}{GSEA scored data (Gene.set, Size, ES, NES, p.value, q.value, gene set)}

\item{title}{string plot title}

\item{gseaParam}{int GSEA parameter value, all gene-level statis are
raised to the power of 'gseaParam' before calculation of GSEA enrichment scores.}
}
\value{
plotly plot object
}
\description{
Generate a plotly-based GSEA Enrichment plot
}
