% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettyRadioButtonsFieldSet.R
\name{prettyRadioButtonsFieldSet}
\alias{prettyRadioButtonsFieldSet}
\title{Create a group of radio buttons with common input}
\usage{
prettyRadioButtonsFieldSet(
  inputId,
  fieldSetData,
  label,
  selected = NULL,
  status = "primary",
  shape = c("round", "square", "curve"),
  outline = FALSE,
  fill = FALSE,
  thick = FALSE,
  animation = NULL,
  icon = NULL,
  plain = FALSE,
  bigger = FALSE,
  inline = FALSE,
  width = NULL
)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{fieldSetData}{A \code{tibble} of values used to build group of radio buttons.}

\item{label}{Display label for the control.}

\item{selected}{The values that should be initially selected,
(if not specified then defaults to the first value).}

\item{status}{Add a class to the radio,
you can use Bootstrap status like 'info', 'primary', 'danger', 'warning' or 'success'.}

\item{shape}{Shape of the radio between \code{square}, \code{curve} and \code{round}.}

\item{outline}{Color also the border of the radio (\code{TRUE} or \code{FALSE}).}

\item{fill}{Fill the radio with color (\code{TRUE} or \code{FALSE}).}

\item{thick}{Make the content inside radio smaller (\code{TRUE} or \code{FALSE}).}

\item{animation}{Add an animation when radio is checked, a value between
\code{smooth}, \code{jelly}, \code{tada}, \code{rotate}, \code{pulse}.}

\item{icon}{Optional, display an icon on the radio, must be an icon created with \code{icon}.}

\item{plain}{Remove the border when radio is checked (\code{TRUE} or \code{FALSE}).}

\item{bigger}{Scale the radio a bit bigger (\code{TRUE} or \code{FALSE}).}

\item{inline}{If \code{TRUE}, render the choices inline (i.e. horizontally).}

\item{width}{The width of the input, e.g. `400px`, or `100\%`.}
}
\value{
returns original dataframe if threshold is met for each group member. If threshold not met, NULL is returned
}
\description{
Create a group of radio buttons with common input
}
