% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runfGSEA.R
\name{runfGSEA}
\alias{runfGSEA}
\title{Run 2-sided fgseaMultilevel analysis (positive and negative) with adjusted pvals}
\usage{
runfGSEA(geneset, ranks, minSize = 15, maxSize = 500, gseaParam = 0, eps = 0)
}
\arguments{
\item{geneset}{list List of gene sets to check.}

\item{ranks}{Named vector of gene-level stats (Gene + "Fold Change" or "rho"). Names should be the same as in 'pathways'}

\item{minSize}{int default 15: Minimal size of a gene set to test. All pathways below the threshold are excluded.}

\item{maxSize}{int default 500: Maximal size of a gene set to test. All pathways above the threshold are excluded.}

\item{gseaParam}{int default 0: GSEA parameter value, all gene-level statis are raised to the power of 'gseaParam' before calculation of GSEA enrichment scores.}

\item{eps}{numeric default 0.0: This parameter sets the boundary for calculating the p value.}
}
\value{
A tibble with GSEA results. Each row corresponds to a tested pathway. The columns are the following
   pathway:  name of the pathway as in 'names(pathway)';
   pval: an enrichment p-value;
   padj – a BH-adjusted p-value;
   log2err – the expected error for the standard deviation of the P-value logarithm.
   ES – enrichment score, same as in Broad GSEA implementation;
   NES – enrichment score normalized to mean enrichment of random samples of the same size;
   size – size of the pathway after removing genes not present in 'names(stats)'.
   leadingEdge – vector with indexes of leading edge genes that drive the enrichment, see http://software.broadinstitute.org/gsea/doc/GSEAUserGuideTEXT.htm#_Running_a_Leading.
}
\description{
Run 2-sided fgseaMultilevel analysis (positive and negative) with adjusted pvals
}
