% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizeByGroup.R
\name{SummarizeByGroup}
\alias{SummarizeByGroup}
\title{Calculate standard summaries across groups for a dataframe}
\usage{
SummarizeByGroup(.data, .summaryVar, ...)
}
\arguments{
\item{.data}{A dataframe}

\item{.summaryVar}{A number to summarize}

\item{...}{groups to summarize within dataframe - can be mutliples.}
}
\value{
dataframe containing with summaries for each group
}
\description{
Calculate standard summaries across groups for a dataframe
}
