% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddSignificanceGroup.R
\name{addSignificanceGroup}
\alias{addSignificanceGroup}
\title{Function to add 3 group significance group label to fold change dataframe}
\usage{
addSignificanceGroup(
  .data,
  foldChangeVar,
  significanceVariable,
  adjustedInd,
  significanceThreshold,
  originalSignificanceThreshold
)
}
\arguments{
\item{.data}{fold change dataframe}

\item{significanceVariable}{name of significance variable column (p-value, q-value, etc)}

\item{adjustedInd}{logical - indicating whether or not a multiple hypothesis correction was applied. When TRUE, "q" will is used. When FALSE "p" is used.}

\item{significanceThreshold}{numeric - value indicating threshold at which an observation in the significanceVariable coulumn is considered statistically significant.}

\item{originalSignificanceThreshold}{numeric - value indicating un-transformed threshold at which an observation in the significanceVariable coulumn is considered statistically significant.}

\item{foldChangeVariable}{name of fold change column}
}
\value{
returns labels and colors for 3 signficance groups (up, down, not significant)
}
\description{
Function to add 3 group significance group label to fold change dataframe
}
