% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyGroupCountThreshold.R
\name{applyGroupCountThreshold}
\alias{applyGroupCountThreshold}
\title{Utility function to check for minimum group membership in a dataframe}
\usage{
applyGroupCountThreshold(.data, .groupVar, groupLabels, threshold = 10)
}
\arguments{
\item{.data}{dataframe containing group column to check}

\item{.groupVar}{column, string - group labels to check}

\item{groupLabels}{vector - labels to check for membership - useful when the source dataframe is missing a required group ex(Sex- Male, Female)}

\item{threshold}{integer - minimum group membership threshold to be met in order to return data}
}
\value{
returns original dataframe if threshold is met for each group member. If threshold not met, NULL is returned
}
\description{
Utility function to check for minimum group membership in a dataframe
}
