% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateFoldChangeByKeyGroup.R
\name{calculateFoldChangeByKeyGroup}
\alias{calculateFoldChangeByKeyGroup}
\title{Calculate fold change for each key value between groups}
\usage{
calculateFoldChangeByKeyGroup(
  .data,
  .key,
  .group,
  .response,
  baselineGroupLabel,
  inf.rm = TRUE,
  ...
)
}
\arguments{
\item{.data}{A dataframe}

\item{.key}{A string or number - key value for dataframe. Statistics will be computed between groups for each key value}

\item{.group}{A string - column indicating group membership - should be binary.}

\item{baselineGroupLabel}{a string - indicating which statisical test to perform by name..}

\item{inf.rm}{logical indicating whether to remove INF fold changes (dividing by 0) - defaults to TRUE}

\item{...}{dots to accomodate additiaon arguments passed when called from parent function}

\item{.value}{A number - numerical value to use with statitical test. Should be log2 transformed values.}
}
\value{
dataframe indicating  fold change (raw and log2) between baseline and comparison for each group by key value
}
\description{
Calculate fold change for each key value between groups
}
