% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatPValue.R
\name{formatPValue}
\alias{formatPValue}
\title{HTML formats p value based on threshold}
\usage{
formatPValue(pvalue, threshold = 0.05)
}
\arguments{
\item{pvalue}{- numeric - p value}

\item{threshold}{- numeric - threshold under which to display formatted p value - defaults to 0.05. For values above threshold, returns "No Significant Difference"}
}
\value{
dataframe with html formatted p value.
}
\description{
HTML formats p value based on threshold
}
