% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatPValue.R
\name{formatPValue}
\alias{formatPValue}
\title{formats significance Variable value value based on threshold and adjustment method}
\usage{
formatPValue(
  significanceVariable,
  adjustmentMethod = "none",
  formatInsignificantValues = FALSE
)
}
\arguments{
\item{significanceVariable}{- numeric - significanceVariable}

\item{adjustmentMethod}{string - multiple hypothesis correction method if any - defaults to "none"}

\item{formatInsignificantValues}{logical - whether to format and return values that are "insignificant" based on threshold}
}
\value{
string - formatted p or q value depending on adjustment method.
}
\description{
formats significance Variable value value based on threshold and adjustment method
}
