% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatPValue.R
\name{formatPValue}
\alias{formatPValue}
\title{HTML formats p value based on threshold}
\usage{
formatPValue(
  p.value,
  p.value.adjustment.method = "none",
  p.value.threshold = 0.05
)
}
\arguments{
\item{p.value}{- numeric - p value}

\item{p.value.adjustment.method}{string - p value adjustment method if any - defaults to "none"}

\item{p.value.threshold}{- numeric - threshold under which to display formatted p value - defaults to 0.05. For values above threshold, returns "No Significant Difference"}
}
\value{
string with html formatted p value.
}
\description{
HTML formats p value based on threshold
}
