% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGroupedStatTestByKeyGroup.R
\name{getGroupedStatTestByKeyGroup}
\alias{getGroupedStatTestByKeyGroup}
\title{function to run stat tests for multiple groups}
\usage{
getGroupedStatTestByKeyGroup(.data, groupVar, ...)
}
\arguments{
\item{.data}{dataframe}

\item{groupVar}{string - column of group values}

\item{...}{additional arguments to be passed to
getStatTestByKeyGroup function}
}
\value{
dataframe containing resulting p.value for each group calculated
}
\description{
function to run stat tests for multiple groups
}
