% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetStatTestByKeyGroup.R
\name{getStatTestByKeyGroup}
\alias{getStatTestByKeyGroup}
\title{Calculate statistical test between 2 groups for each key value in a dataframe}
\usage{
getStatTestByKeyGroup(
  .data,
  id,
  key,
  response,
  independentVariable,
  baselineLabel,
  testMethod,
  ...
)
}
\arguments{
\item{.data}{A dataframe}

\item{...}{dots - accomodate additional arguments when function is called as parent router}

\item{.id}{A string or numeric column - represents a unique observation within each key/group combination. PersonId, LabId, etc}

\item{.key}{A string or numeric column - key value for dataframe. Statistics will be computed between groups for each key value}

\item{.group}{A string column indicating group membership - should be binary.}

\item{.value}{A numeric column - numerical value to use with statitical test between groups.}

\item{method}{a string - indicating which statisical test to perform. One of either ks.test, t.test, or wilcox.test}
}
\value{
dataframe containing resulting p.values for each key/group stat test
}
\description{
Calculate statistical test between 2 groups for each key value in a dataframe
}
