% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetStatTestByKeyGroup.R
\name{GetStatTestByKeyGroup}
\alias{GetStatTestByKeyGroup}
\title{Calculate statistical test across 2 groups for each key value in a dataframe}
\usage{
GetStatTestByKeyGroup(
  .data,
  .id,
  .key,
  .group,
  .value,
  .method,
  .addLog10 = TRUE
)
}
\arguments{
\item{.data}{A dataframe}

\item{.id}{A string or number - represents a unique observation within each key/group combination. PersonId, LabId, etc}

\item{.key}{A string or number - key value for dataframe. Statistics will be computed between groups for each key value}

\item{.group}{A string - column indicating group membership - should be binary.}

\item{.value}{A number - numerical value to use with statitical test.}

\item{.method}{a string - indicating which statisical test to perform by name..}

\item{.addLog10}{boolean - whether or not to add a -log10 transfromation of resulting p.values}
}
\value{
dataframe containing resulting p.values for each key/group stat test
}
\description{
Calculate statistical test across 2 groups for each key value in a dataframe
}
