% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVolcanoPlot.R
\name{getVolcanoPlot}
\alias{getVolcanoPlot}
\title{Plotly function to generate standard volcano plot from fold change data}
\usage{
getVolcanoPlot(.data, .xvar, .yvar, .significanceGroup, .text, .key, plotName)
}
\arguments{
\item{.data}{dataframe containing fold change data}

\item{.xvar}{fold change variable}

\item{.yvar}{p value variable}

\item{.significanceGroup}{significance group column}

\item{.text}{column containing text values to show in tooltip}

\item{.key}{- key column in fold change data - will be used for capuring select events}

\item{plotName}{- name to attribute to plot (used for tracking clicks, events, etc)}
}
\value{
returns list of lists of plotly annotation objects
         -- up regulated arror at top of plot (with up regulated text)
         -- down regulated arrow at the top of plot (
         -- dp value threshold text with "p(a) > threshold" text and up arrow
         -- if a value is chosen, adds in arrow annotation.
}
\description{
Plotly function to generate standard volcano plot from fold change data
}
