% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVolcanoPlot.R
\name{getVolcanoPlot}
\alias{getVolcanoPlot}
\title{Plotly function to generate standard volcano plot from fold change data}
\usage{
getVolcanoPlot(
  .data,
  foldChangeVariable,
  significanceVariable,
  significanceGroup,
  text,
  key,
  plotName,
  color,
  shape = "circle"
)
}
\arguments{
\item{.data}{dataframe containing fold change data}

\item{foldChangeVariable}{fold change variable}

\item{significanceVariable}{p value variable}

\item{significanceGroup}{significance group column - each group
results in a separate trace}

\item{text}{column containing text values to show in tooltip}

\item{key}{- key column in fold change data - will be used for
capuring select events}

\item{plotName}{- name to attribute to plot
(used for tracking clicks, events, etc)}

\item{color}{- column with color to use for each row / group in dataset -
should be unique per significance group / trace}

\item{shape}{- optional shape - should be passed as a nammed
column in source data.}
}
\value{
returns Plotly scatter plot showing fold change vs p value
colored by significance group
}
\description{
Plotly function to generate standard volcano plot from fold change data
}
