% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVolcanoPlot.R
\name{getVolcanoPlot}
\alias{getVolcanoPlot}
\title{Plotly function to generate standard volcano plot from fold change data}
\usage{
getVolcanoPlot(
  .data,
  foldChangeVar,
  pValueVar,
  significanceGroup,
  text,
  key,
  plotName
)
}
\arguments{
\item{.data}{dataframe containing fold change data}

\item{significanceGroup}{significance group column}

\item{text}{column containing text values to show in tooltip}

\item{key}{- key column in fold change data - will be used for capuring select events}

\item{plotName}{- name to attribute to plot (used for tracking clicks, events, etc)}

\item{xvar}{fold change variable}

\item{yvar}{p value variable}
}
\value{
returns Plotly scatter plot showing fold change vs p value colored by significance group
}
\description{
Plotly function to generate standard volcano plot from fold change data
}
