% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizeByGroup.R
\name{summarizeByGroup}
\alias{summarizeByGroup}
\title{Calculate standard summaries across groups for a dataframe}
\usage{
summarizeByGroup(.data, .summaryVar, ..., na.rm = TRUE)
}
\arguments{
\item{.data}{A dataframe}

\item{.summaryVar}{A number to summarize}

\item{...}{group variables to summarize within dataframe - can be mutliples.}

\item{na.rm}{boolean - indicates whether NA values should be stripped 
from summary calculation. Defaults to TRUE}
}
\value{
dataframe containing with summaries for each group
}
\description{
Calculate standard summaries across groups for a dataframe
}
